Move loader to component.
