+++ src/main/static/src/components/common/icons/Loader.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Loader = () => (<div className="oph-spinner">
+    <div className="oph-bounce oph-bounce1"/>
+    <div className="oph-bounce oph-bounce2"/>
+    <div className="oph-bounce oph-bounce3"/>
+</div>);
+
+export default Loader;
