+++ src/main/static/src/components/henkilo/VirkailijaViewContainer.js
@@ -10,7 +10,7 @@ import {
 } from "../../actions/koodisto.actions";
 import {updateNavigation} from "../../actions/navigation.actions";
 import {virkailijaNavi} from "../../configuration/navigationconfigurations";
-import AbstractViewContainer from "./AbstractViewContainer";
+import AbstractViewContainer from "../../containers/henkilo/AbstractViewContainer";
 import {
     addKayttooikeusToHenkilo,
     fetchAllKayttooikeusAnomusForHenkilo,
