+++ src/main/static/src/components/henkilo/VirkailijaViewPage.js
@@ -7,6 +7,7 @@ import HenkiloViewExistingKayttooikeus from "../common/henkilo/HenkiloViewExisti
 import HenkiloViewExpiredKayttooikeus from "../common/henkilo/HenkiloViewExpiredKayttooikeus";
 import HenkiloViewOpenKayttooikeusanomus from "../common/henkilo/HenkiloViewOpenKayttooikeusanomus";
 import HenkiloViewCreateKayttooikeus from "../common/henkilo/HenkiloViewCreateKayttooikeus";
+import Loader from "../common/icons/Loader";
 
 class VirkailijaViewPage extends React.Component {
     constructor(props) {
@@ -20,7 +21,7 @@ class VirkailijaViewPage extends React.Component {
                 <div className="wrapper">
                     {
                         this.props.isUserContentLoading()
-                            ? AbstractViewContainer.createLoader()
+                            ? <Loader />
                             : <HenkiloViewUserContent {...this.props} readOnly={true} locale={this.props.locale} showPassive={false}
                                                       basicInfo={this.props.createBasicInfo}
                                                       basicInfo2={this.props.createBasicInfo2}
@@ -32,14 +33,14 @@ class VirkailijaViewPage extends React.Component {
                 <div className="wrapper">
                     {
                         this.props.isOrganisationContentLoading()
-                            ? AbstractViewContainer.createLoader()
+                            ? <Loader />
                             : <HenkiloViewOrganisationContent {...this.props} readOnly={true} locale={this.props.locale} />
                     }
                 </div>
                 <div className="wrapper">
                     {
                         this.props.kayttooikeus.kayttooikeusLoading
-                            ? AbstractViewContainer.createLoader()
+                            ? <Loader />
                             : <HenkiloViewExistingKayttooikeus {...this.props}
                                                                ref={(ref) => this.existingKayttooikeusRef = ref} />
                     }
@@ -47,21 +48,21 @@ class VirkailijaViewPage extends React.Component {
                 <div className="wrapper">
                     {
                         this.props.kayttooikeus.kayttooikeusAnomusLoading
-                            ? AbstractViewContainer.createLoader()
+                            ? <Loader />
                             : <HenkiloViewOpenKayttooikeusanomus {...this.props} />
                     }
                 </div>
                 <div className="wrapper">
                     {
                         this.props.kayttooikeus.kayttooikeusLoading
-                            ? AbstractViewContainer.createLoader()
+                            ? <Loader />
                             : <HenkiloViewExpiredKayttooikeus {...this.props} />
                     }
                 </div>
                 <div className="wrapper">
                     {
                         this.props.henkilo.henkiloOrganisaatiosLoading
-                            ? AbstractViewContainer.createLoader()
+                            ? <Loader />
                             : <HenkiloViewCreateKayttooikeus {...this.props}
                                                              existingKayttooikeusRef={this.existingKayttooikeusRef} />
                     }
